refactor(*): massive restructure of app
Found I wasn't following a very good, scalable, file structure convention with express or angular, so I've restructured everything.
It's early enough that this won't be too painful, but I still have to rename most of the angular modules.

This likely breaks the entire angular app, but the express side should be repathed.